refactor: replace original effects UI with multiband EQ
