+++ src/components/recorder/eq-response-graph.tsx
@@ -2,11 +2,12 @@ import {
   type PointerEvent as ReactPointerEvent,
   useCallback,
   useEffectEvent,
+  useRef,
 } from "react";
 import {
   calculateBiquadEqCoefficients,
   calculateBiquadEqResponse,
-  type EqParameters,
+  type MultibandEqBand,
 } from "../../lib/dsp/biquad-eq";
 import { clamp, dbToGain, gainToDb } from "../../lib/music";
 import { EQ_CONTROL_LIMITS } from "./eq-control-limits";
@@ -14,75 +15,91 @@ import { EQ_CONTROL_LIMITS } from "./eq-control-limits";
 const GRAPH_SAMPLE_RATE = 48000;
 const FREQUENCY_TICKS = [20, 100, 1000, 10000, 20000];
 const GAIN_TICKS = [-18, -12, -6, 0, 6, 12, 18];
+export const EQ_BAND_COLORS = [
+  "#60a5fa",
+  "#f472b6",
+  "#fbbf24",
+  "#34d399",
+  "#a78bfa",
+  "#fb7185",
+  "#22d3ee",
+  "#a3e635",
+];
 
 export function EqResponseGraph({
-  eq,
-  onChange,
+  bands,
+  selectedBandId,
+  bypass,
+  onSelectBand,
+  onBandChange,
 }: {
-  eq: EqParameters;
-  onChange: (update: Partial<EqParameters>) => void;
+  bands: MultibandEqBand[];
+  selectedBandId?: string;
+  bypass: boolean;
+  onSelectBand: (id: string) => void;
+  onBandChange: (
+    id: string,
+    update: Partial<Omit<MultibandEqBand, "id">>,
+  ) => void;
 }) {
-  // The plot uses normalized log-frequency and gain coordinates from 0 to 1.
-  const coefficients = calculateBiquadEqCoefficients({
-    sampleRate: GRAPH_SAMPLE_RATE,
-    frequency: eq.frequency,
-    gain: eq.gain,
-    q: eq.q,
-  });
-  const responsePath = Array.from({ length: 161 }, (_, index) => {
-    const x = index / 160;
-    const frequency = graphXToFrequency(x);
-    const gainDb = gainToDb(
-      calculateBiquadEqResponse({
-        coefficients,
-        sampleRate: GRAPH_SAMPLE_RATE,
-        frequency,
-      }),
-    );
-    return `${index === 0 ? "M" : "L"}${x.toFixed(5)},${gainDbToGraphY(gainDb).toFixed(5)}`;
-  }).join(" ");
+  const plotElementRef = useRef<HTMLDivElement>(null);
+  const bandResponses = bands.map((band) => ({
+    band,
+    path: createResponsePath([band]),
+  }));
+  const combinedPath = createResponsePath(bands.filter((band) => !band.bypass));
 
-  // Pointer edits map the plot position back to frequency and gain.
-  const updateFromPointer = (event: ReactPointerEvent<HTMLDivElement>) => {
-    const bounds = event.currentTarget.getBoundingClientRect();
+  const updateFromPointer = (
+    event: ReactPointerEvent<HTMLButtonElement>,
+    id: string,
+  ) => {
+    const bounds = plotElementRef.current!.getBoundingClientRect();
     const x = clamp((event.clientX - bounds.left) / bounds.width, 0, 1);
     const y = clamp((event.clientY - bounds.top) / bounds.height, 0, 1);
     const frequency = Math.round(graphXToFrequency(x));
     const step = EQ_CONTROL_LIMITS.gainDb.step;
     const gainDb = Math.round(graphYToGainDb(y) / step) * step;
-    onChange({ frequency, gain: dbToGain(gainDb) });
+    onBandChange(id, { frequency, gain: dbToGain(gainDb) });
   };
 
-  // Wheel gestures adjust Q independently of the point position.
   const handleWheel = useEffectEvent((event: WheelEvent) => {
-    if (event.ctrlKey || event.metaKey || event.deltaY === 0) {
+    if (
+      !selectedBandId ||
+      event.ctrlKey ||
+      event.metaKey ||
+      event.deltaY === 0
+    ) {
+      return;
+    }
+    const band = bands.find((entry) => entry.id === selectedBandId);
+    if (!band) {
       return;
     }
     event.preventDefault();
-    const q = clamp(
-      eq.q * Math.exp(event.deltaY * 0.002),
-      EQ_CONTROL_LIMITS.q.min,
-      EQ_CONTROL_LIMITS.q.max,
-    );
-    onChange({ q });
+    onBandChange(selectedBandId, {
+      q: clamp(
+        band.q * Math.exp(event.deltaY * 0.002),
+        EQ_CONTROL_LIMITS.q.min,
+        EQ_CONTROL_LIMITS.q.max,
+      ),
+    });
   });
-  const graphRef = useCallback(
-    (graph: HTMLDivElement | null) => {
-      if (!graph) {
+  const plotRef = useCallback(
+    (plot: HTMLDivElement | null) => {
+      plotElementRef.current = plot;
+      if (!plot) {
         return;
       }
-      // A non-passive listener consumes Q gestures without scrolling the panel.
-      graph.addEventListener("wheel", handleWheel, { passive: false });
-      return () => graph.removeEventListener("wheel", handleWheel);
+      plot.addEventListener("wheel", handleWheel, { passive: false });
+      return () => plot.removeEventListener("wheel", handleWheel);
     },
     [handleWheel],
   );
 
-  // Axis labels sit outside the measured plot, so layout does not affect gestures.
   return (
     <div
       data-testid="eq-response-graph"
-      aria-label="EQ response graph"
+      aria-label="Multiband EQ response graph"
       className="grid aspect-[8/5] w-full grid-cols-[34px_1fr] grid-rows-[1fr_22px] rounded border border-neutral-700 bg-neutral-900 pt-2 pr-2 text-[9px] text-neutral-500 select-none"
     >
       <div className="relative">
@@ -96,25 +113,11 @@ export function EqResponseGraph({
           </span>
         ))}
       </div>
-      <div
-        ref={graphRef}
-        className="relative min-h-0 min-w-0 touch-none cursor-crosshair"
-        onPointerDown={(event) => {
-          if (event.button === 0) {
-            event.currentTarget.setPointerCapture(event.pointerId);
-            updateFromPointer(event);
-          }
-        }}
-        onPointerMove={(event) => {
-          if (event.currentTarget.hasPointerCapture(event.pointerId)) {
-            updateFromPointer(event);
-          }
-        }}
-      >
+      <div ref={plotRef} className="relative min-h-0 min-w-0 touch-none">
         <svg
           viewBox="0 0 1 1"
           preserveAspectRatio="none"
-          className="pointer-events-none absolute size-full overflow-visible"
+          className="pointer-events-none absolute size-full overflow-hidden"
           aria-hidden="true"
         >
           {GAIN_TICKS.map((gainDb) => (
@@ -143,23 +146,65 @@ export function EqResponseGraph({
               vectorEffect="non-scaling-stroke"
             />
           ))}
+          {bandResponses.map(({ band, path }, index) => (
+            <path
+              key={band.id}
+              data-testid="eq-band-curve"
+              d={path}
+              fill="none"
+              stroke={EQ_BAND_COLORS[index]}
+              strokeOpacity={
+                band.bypass ? 0.18 : band.id === selectedBandId ? 0.8 : 0.35
+              }
+              strokeWidth={band.id === selectedBandId ? 1.5 : 1}
+              strokeDasharray={band.bypass ? "3 3" : undefined}
+              vectorEffect="non-scaling-stroke"
+            />
+          ))}
           <path
-            data-testid="eq-response-curve"
-            d={responsePath}
+            data-testid="eq-combined-curve"
+            d={combinedPath}
             fill="none"
-            className={eq.bypass ? "stroke-blue-400/35" : "stroke-blue-400"}
-            strokeWidth={2}
+            className="stroke-neutral-100"
+            strokeOpacity={bypass ? 0.25 : 0.9}
+            strokeWidth={2.5}
             vectorEffect="non-scaling-stroke"
           />
         </svg>
-        <span
-          data-testid="eq-response-point"
-          className="pointer-events-none absolute size-3 -translate-x-1/2 -translate-y-1/2 rounded-full border-2 border-blue-300 bg-neutral-900"
-          style={{
-            left: `${frequencyToGraphX(eq.frequency) * 100}%`,
-            top: `${gainDbToGraphY(gainToDb(eq.gain)) * 100}%`,
-          }}
-        />
+        {bands.map((band, index) => {
+          const selected = band.id === selectedBandId;
+          return (
+            <button
+              key={band.id}
+              type="button"
+              data-testid="eq-response-point"
+              aria-label={`Select and edit band ${index + 1}`}
+              aria-pressed={selected}
+              onPointerDown={(event) => {
+                if (event.button !== 0) {
+                  return;
+                }
+                event.currentTarget.setPointerCapture(event.pointerId);
+                onSelectBand(band.id);
+                updateFromPointer(event, band.id);
+              }}
+              onPointerMove={(event) => {
+                if (event.currentTarget.hasPointerCapture(event.pointerId)) {
+                  updateFromPointer(event, band.id);
+                }
+              }}
+              className="absolute flex size-5 -translate-x-1/2 -translate-y-1/2 cursor-grab items-center justify-center rounded-full border-2 bg-neutral-900 text-[9px] font-semibold text-neutral-100 shadow-sm outline-none hover:scale-110 focus-visible:ring-2 focus-visible:ring-white active:cursor-grabbing aria-pressed:scale-110"
+              style={{
+                left: `${frequencyToGraphX(band.frequency) * 100}%`,
+                top: `${gainDbToGraphY(gainToDb(band.gain)) * 100}%`,
+                borderColor: EQ_BAND_COLORS[index],
+                opacity: band.bypass ? 0.45 : 1,
+              }}
+            >
+              {index + 1}
+            </button>
+          );
+        })}
       </div>
       <div className="relative col-start-2">
         {FREQUENCY_TICKS.map((frequency) => (
@@ -174,6 +219,34 @@ export function EqResponseGraph({
       </div>
     </div>
   );
+}
+
+function createResponsePath(bands: MultibandEqBand[]): string {
+  const coefficients = bands.map((band) =>
+    calculateBiquadEqCoefficients({
+      sampleRate: GRAPH_SAMPLE_RATE,
+      frequency: band.frequency,
+      gain: band.gain,
+      q: band.q,
+    }),
+  );
+  return Array.from({ length: 161 }, (_, index) => {
+    const x = index / 160;
+    const frequency = graphXToFrequency(x);
+    const gainDb = coefficients.reduce(
+      (sum, entry) =>
+        sum +
+        gainToDb(
+          calculateBiquadEqResponse({
+            coefficients: entry,
+            sampleRate: GRAPH_SAMPLE_RATE,
+            frequency,
+          }),
+        ),
+      0,
+    );
+    return `${index === 0 ? "M" : "L"}${x.toFixed(5)},${gainDbToGraphY(gainDb).toFixed(5)}`;
+  }).join(" ");
 }
 
 function frequencyToGraphX(frequency: number): number {
